requirements.txt
Flask==2.3.3
Jinja2==3.1.3
Werkzeug==2.3.7
openpyxl==3.1.2  # for Excel export
pandas
flask_sqlalchemy
fpdf
weasyprint
reportlab
pyinstaller
pywebview
waitress
psycopg2-binary
python-dotenv